2026.03.06 08:53 (Cloud Seed Propagation for Deterministic Cloud Layout Rebuilds)
- Added explicit integer-seed propagation semantics for the cloud-layout control path.
The cloud placement pipeline was restored to a type-stable configuration flow in which the pause-menu slider, the Qt signal-slot boundary, the viewport runtime state, and the renderer-side cloud field all treat the seed as an integer-valued control parameter rather than an implicitly widened scalar. This is materially significant because the cloud generator is designed as a deterministic discrete-layout system whose spatial arrangement is keyed by integer seed identity. By re-establishing that invariant at the UI emission boundary, the seed change once again becomes a semantically valid invalidation event for cached cloud placement, thereby allowing the cloud field to rebuild against a genuinely new deterministic state instead of propagating an ill-typed value through the control chain.

- Changed the Cloud seed emission path from floating-point delivery to strict integer delivery.
The defect originated in the pause overlay’s cloud-seed update handler, where an int-declared Qt signal was being emitted with a float payload. Although the surrounding runtime path nominally expected integral seed values, this mismatch weakened the configuration contract at the exact point where user input is serialized into renderer-facing state. The fix therefore replaced floating-point emission with strict integer emission, aligning the produced payload with the declared signal type and with the downstream seed consumers. As a result, adjusting Cloud seed now reliably reaches GLViewportWidget, updates the runtime state as an integer seed, invokes GLRenderer.set_cloud_seed(...) without type ambiguity, and correctly triggers CloudPass and CloudField cache invalidation so that cloud placement is recomputed and visibly changes in response to slider manipulation.

+++ src/maiming/presentation/widgets/overlays/pause_overlay.py
@@ -351,4 +351,4 @@ class PauseOverlay(QWidget):
     def _on_cloud_seed(self, v: int) -> None:
         sv = int(v)
         self._lbl_cloud_seed.setText(f"Cloud seed: {sv}")
-        self.cloud_seed_changed.emit(float(sv))
+        self.cloud_seed_changed.emit(int(sv))
